Refactors attachable repair migration
